+++ src/main/java/rzk/wirelessredstone/util/LangKeys.java
@@ -13,4 +13,10 @@ public class LangKeys
 	public static final String TOOLTIP_ON = "tooltip.wirelessredstone.on.name";
 
 	public static final String COMMAND_CLEAR = "command.wirelessredstone.clear.usage";
+	public static final String COMMAND_CLEAR_SUCCESS = "command.wirelessredstone.clear.success";
+	public static final String COMMAND_CLEAR_SUCCESS_ALL = "command.wirelessredstone.clear.success.all";
+
+	public static final String MESSAGE_ACTIVE_TRANSMITTERS = "message.wirelessredstone.active.transmitters";
+	public static final String MESSAGE_NO_TRANSMITTERS = "message.wirelessredstone.no.transmitters";
+	public static final String MESSAGE_TELEPORT = "message.wirelessredstone.teleport";
 }